Fix WatermarkService type casting issue
- Fixed parameter type casting in imagecopymerge function
- Cast opacity calculation result to int as expected by function
- Improves code quality and resolves PHPStan error

+++ app/Services/WatermarkService.php
@@ -280,7 +280,7 @@ final class WatermarkService
             0,
             $watermarkWidth,
             $watermarkHeight,
-            (float) $this->config['opacity'] * 100
+            (int) ((float) $this->config['opacity'] * 100)
         );
 
         // Clean up watermark resource
